Change coded_block_pattern to uint32 to make it easier to read in debugger.

+++ src/mpeg1.cpp
@@ -468,6 +468,6 @@ size_t mpeg1::calc_dct_zz_zero(size_t dc_size, size_t dc_differential) {
     }
 }
 
-bool mpeg1::check_cbp(int8_t coded_block_pattern, size_t index) {
+bool mpeg1::check_cbp(uint32_t coded_block_pattern, size_t index) {
     return coded_block_pattern & (1 << (5 - index));
 }
